Add files via upload
nawk and gawk versions that take a .ics (iCalendar) file and turn it into something calcurse likes. 
This is primarily a readability improvement for the user.

TODO: ics description does *not* get dumped into a note, I don't know how to make awk calculate the sha1 hash for the note name.